DB tables have plural names and categories are normal

+++ RR.Common/Models/Receipt.cs
@@ -4,7 +4,7 @@ public class Receipt
 {
     public Vendor Vendor { get; set; }
     public List<ReceiptItem> Items { get; set; } = [];
-    public long Total => Items.Sum(i => i.Price);
+    public double Total => Items.Sum(i => i.Price);
     public string? ReceiptId { get; set; }
     public DateTime? DateTime { get; set; }
 }
